Replace unnecessary custom components
