+++ airbyte-integrations/connectors/source-square/source_square/manifest.yaml
@@ -1,4 +1,4 @@
-version: "0.29.0"
+version: "0.81.4"
 
 definitions:
   schema_loader:
@@ -327,26 +327,21 @@ definitions:
           field_path: ["loyalty_accounts"]
 
   orders_stream:
+    # ToDo: Improve the efficiency of this stream by grouping location IDs into batches of 10.
     $ref: "#/definitions/base_stream_page_json_limit"
     $parameters:
       name: "orders"
       primary_key: "id"
       path: "/orders/search"
     incremental_sync:
-      type: CustomIncrementalSync
-      class_name: source_square.components.SquareSubstreamIncrementalSync
+      type: DatetimeBasedCursor
       start_datetime:
         datetime: "{{ format_datetime(config['start_date'], '%Y-%m-%dT%H:%M:%S.%fZ') }}"
-        datetime_format: "%Y-%m-%dT%H:%M:%S.%fZ"
-      end_datetime:
-        datetime: "{{ now_utc().strftime('%Y-%m-%dT%H:%M:%S.%fZ') }}"
         datetime_format: "%Y-%m-%dT%H:%M:%S.%fZ"
       step: P30D
       datetime_format: "%Y-%m-%dT%H:%M:%S.%fZ"
       cursor_granularity: "PT0.000001S"
-      parent_stream: "#/definitions/locations_stream"
       cursor_field: "updated_at"
-      parent_key: "id"
     retriever:
       $ref: "#/definitions/base_stream_page_json_limit/retriever"
       requester:
@@ -354,6 +349,20 @@ definitions:
         http_method: "POST"
         request_body_json:
           limit: "{{ 500 }}"
+          filter:
+            date_time_filter:
+              updated_at:
+                start_at: "{{ stream_slice.start_time }}"
+          sort:
+            sort_field: "UPDATED_AT"
+            sort_order: "ASC"
+          location_ids: "{{ [ stream_slice.location_ids ] }}"
+      partition_router:
+        type: "SubstreamPartitionRouter"
+        parent_stream_configs:
+          - stream: "#/definitions/locations_stream"
+            parent_key: "id"
+            partition_field: "location_ids"
 
   bank_accounts_stream:
     $ref: "#/definitions/base_stream"
